Use the ProductDAO in the JdbcExample class
Here we are not removing the connection object from the try/catch with resources because we want the resource closed when the block of code has ended. Depending on how you are architecting an application, you may want to keep the connection alive to use it across the application, or you may want to close it manually inside the DAO right after pulling out the data.